+++ CHANGELOG.md
@@ -2,5 +2,7 @@
 
 ## Unreleased
 
+- (JA) リリース時にsketch.yamlの`dir`をライブラリ名(バージョン)へ置換し、Gitにはステージしないように変更
+- (EN) Updated release flow to swap `dir` entries in sketch.yaml to library name (version) without staging them in Git
 - (JA) 初期コミット
 - (EN) Initial commit